feat: Created ZoneEntryFacade (name pending) which is a factory facade linking the classes that modify values involved in calculating zones, created classes to detect zone changes, refactored a lot of the pef entry parts, and created personal best entry parts

+++ Starter Application/app/src/main/java/com/example/smartAir/triaging/RecentHistoryEntry.java
@@ -44,7 +44,7 @@ public class RecentHistoryEntry extends Fragment {
             public void onTextChanged(CharSequence s, int start, int before, int count) {
 
                 if (!s.isEmpty()) {
-                    triageController.setPEF(Double.parseDouble(s.toString()));
+                    triageController.setPEF(Float.parseFloat(s.toString()));
                 } else {
                     triageController.setPEF(null);
                 }
